Added creator ID references to posts

+++ backend/routes/posts.js
@@ -38,6 +38,7 @@ router.post(
       title: req.body.title,
       content: req.body.content,
       imagePath: serverUrl + "/images/" + req.file.filename,
+      creator: req.userData.userId,
     });
     post.save().then((createdPost) => {
       res.status(201).json({
@@ -81,9 +82,13 @@ router.get("", (req, res, next) => {
 
 // DELETE a post
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then((result) => {
-      res.status(200).json({ message: "Record deleted" });
+      if (result.n > 0) {
+        res.status(200).json({ message: "Delete successful" });
+      } else {
+        res.status(401).json({ message: "Not authorized" });
+      }
     })
     .catch((e) => {
       return res.status(500).send();
@@ -106,9 +111,17 @@ router.put(
       title: req.body.title,
       content: req.body.content,
       imagePath: imagePath,
+      creator: req.userData.userId,
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({ message: "Update successful" });
+    Post.updateOne(
+      { _id: req.params.id, creator: req.userData.userId },
+      post
+    ).then((result) => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful" });
+      } else {
+        res.status(401).json({ message: "Not authorized" });
+      }
     });
   }
 );
